fix : login footer 로 명칭 변경
