+++ src/screens/login/modules/index.ts
@@ -1,4 +1,4 @@
-export * from './login-confirm-button';
+export * from './login-footer';
 export * from './login-id-input';
 export * from './login-pw-input';
 export * from './login-header';
